@@ -2,42 +2,18 @@
 
 namespace Contributte\Invoice\Data;
 
-use Nette\SmartObject;
-
-class Account
+class Account implements IAccount
 {
 
-	use SmartObject;
-
-	/** @var string */
-	private $accountNumber;
-
-	/** @var string|null */
-	private $iBan;
-
-	/** @var string|null */
-	private $swift;
-
-	public function __construct(string $accountNumber, ?string $iBan = null, ?string $swift = null)
+	public function __construct(
+		private ?string $iban = null,
+	)
 	{
-		$this->accountNumber = $accountNumber;
-		$this->iBan = $iBan;
-		$this->swift = $swift;
 	}
 
-	public function getAccountNumber(): string
+	public function getIban(): ?string
 	{
-		return $this->accountNumber;
-	}
-
-	public function getIBan(): ?string
-	{
-		return $this->iBan;
-	}
-
-	public function getSwift(): ?string
-	{
-		return $this->swift;
+		return $this->iban;
 	}
 
 }
