@@ -2,19 +2,15 @@
 
 namespace Contributte\Invoice\DI;
 
+use Contributte\Invoice\Data\Account;
 use Contributte\Invoice\Data\Company;
-use Contributte\Invoice\Formatter;
-use Contributte\Invoice\IFormatter;
 use Contributte\Invoice\Invoice;
-use Contributte\Invoice\ITranslator;
-use Contributte\Invoice\Renderers\IRenderer;
-use Contributte\Invoice\Renderers\PDFRenderer;
-use Contributte\Invoice\Templates\DefaultTemplate;
-use Contributte\Invoice\Templates\ITemplate;
-use Contributte\Invoice\Translator;
+use Contributte\Invoice\Provider\InvoiceDataProvider;
 use Nette\DI\CompilerExtension;
+use Nette\DI\Definitions\Statement;
 use Nette\Schema\Expect;
 use Nette\Schema\Schema;
+use stdClass;
 
 class InvoiceExtension extends CompilerExtension
 {
@@ -28,41 +24,67 @@ class InvoiceExtension extends CompilerExtension
 				'address' => Expect::string()->required(),
 				'zip' => Expect::string()->required(),
 				'country' => Expect::string()->required(),
-				'tin' => Expect::string(),
-				'vaTin' => Expect::string(),
-				'hasTax' => Expect::bool(false),
+				'vatNumber' => Expect::string(),
+				'id' => Expect::string(),
 			]),
-			'lang' => Expect::string('en'),
-			'template' => Expect::string(DefaultTemplate::class),
+			'accounts' => Expect::arrayOf(Expect::structure([
+				'iban' => Expect::string(),
+			])),
 		]);
 	}
 
 	public function loadConfiguration(): void
 	{
 		$builder = $this->getContainerBuilder();
-		$config = (array) $this->getConfig();
 
-		$builder->addDefinition($this->prefix('company'))
-			->setFactory(Company::class, array_values((array) $config['company']));
-
-		$builder->addDefinition($this->prefix('template'))
-			->setFactory($config['template'])
-			->setType(ITemplate::class);
-
-		$builder->addDefinition($this->prefix('renderer'))
-			->setFactory(PDFRenderer::class)
-			->setType(IRenderer::class);
-
-		$builder->addDefinition($this->prefix('translation'))
-			->setType(ITranslator::class)
-			->setFactory(Translator::class, [$config['lang']]);
-
-		$builder->addDefinition($this->prefix('formatter'))
-			->setType(IFormatter::class)
-			->setFactory(Formatter::class, [$config['lang']]);
+		$builder->addDefinition($this->prefix('dataProvider'))
+			->setFactory(InvoiceDataProvider::class, [
+				$this->getCompany(),
+				$this->getAccounts(),
+			]);
 
 		$builder->addDefinition($this->prefix('invoice'))
 			->setFactory(Invoice::class);
+	}
+
+	private function getCompany(): ?Statement
+	{
+		/** @var stdClass $config */
+		$config = $this->getConfig();
+
+		/** @var stdClass $company */
+		$company = $config->company;
+
+		if (!$company->name) {
+			return null;
+		}
+
+		return new Statement(Company::class, [
+			$company->name,
+			$company->town,
+			$company->address,
+			$company->zip,
+			$company->country,
+			$company->vatNumber,
+			$company->id,
+		]);
+	}
+
+	/**
+	 * @return Statement[]
+	 */
+	private function getAccounts(): array
+	{
+		/** @var stdClass $config */
+		$config = $this->getConfig();
+
+		$accounts = [];
+		/** @var stdClass $account */
+		foreach ($config->accounts as $account) {
+			$accounts[] = new Statement(Account::class, [$account->iban]);
+		}
+
+		return $accounts;
 	}
 
 }
