@@ -5,36 +5,96 @@ namespace Contributte\Invoice\Preview;
 use Contributte\Invoice\Data\Account;
 use Contributte\Invoice\Data\Company;
 use Contributte\Invoice\Data\Customer;
+use Contributte\Invoice\Data\International\Czech\CzechAccount;
+use Contributte\Invoice\Data\International\Czech\CzechPaymentInformation;
+use Contributte\Invoice\Data\IOrder;
+use Contributte\Invoice\Data\Item;
 use Contributte\Invoice\Data\Order;
 use Contributte\Invoice\Data\PaymentInformation;
+use Contributte\Invoice\Data\Timestamps;
 use DateTime;
 
 final class PreviewFactory
 {
 
-	public static function createCompany(): Company
+	public static function createOrder(?int $itemCount = null): IOrder
 	{
-		return new Company('Contributte', 'Prague', 'U haldy', '110 00', 'Czech Republic', '08304431', 'CZ08304431');
-	}
+		$order = new Order(
+			date('Y') . '0001',
+			'$ 15.000,00',
+			new Company('Contributte', 'Prague', 'U haldy', '110 00', 'Czech Republic', 'CZ08304431', '08304431'),
+			new Customer('John Doe', 'Los Angeles', 'Cavetown', '720 55', 'USA', 'CZ08304431', '08304431'),
+			new PaymentInformation(
+				[new Account('CZ4808000000002353462013')],
+			),
+			new Timestamps(
+				(new DateTime())->format('Y-m-d'),
+				(new DateTime('+ 1 week'))->format('Y-m-d'),
+			),
+		);
 
-	public static function createCustomer(): Customer
-	{
-		return new Customer('John Doe', 'Los Angeles', 'Cavetown', '720 55', 'USA', '08304431', 'CZ08304431');
-	}
-
-	public static function createOrder(): Order
-	{
-		$account = new Account('2353462013/0800', 'CZ4808000000002353462013', 'GIGACZPX');
-		$paymentInfo = new PaymentInformation('$', '0123456789', '1234', 0.21);
-
-		$order = new Order(date('Y') . '0001', new DateTime('+ 7 days'), $account, $paymentInfo);
-		$order->addItem('Logitech G700s Rechargeable Gaming Mouse', 1790, 4);
-		$order->addItem('ASUS Z380KL 8" - 16GB, LTE, bílá', 6490, 1);
-		$order->addItem('Philips 48PFS6909 - 121cm', 13990, 1);
-		$order->addItem('HP Deskjet 3545 Advantage', 1799, 1);
-		$order->addItem('LG 105UC9V - 266cm', 11599, 2);
+		self::addItems($order, $itemCount);
 
 		return $order;
+	}
+
+	public static function createCzechOrder(?int $itemCount = null): IOrder
+	{
+		$order = new Order(
+			date('Y') . '0001',
+			'15.000,00 Kč',
+			new Company('Contributte', 'Prague', 'U haldy', '110 00', 'Czech Republic', 'CZ08304431', '08304431'),
+			new Customer('John Doe', 'Los Angeles', 'Cavetown', '720 55', 'USA', 'CZ08304431', '08304431'),
+			new CzechPaymentInformation(
+				[new CzechAccount('2353462013/0800', 'CZ4808000000002353462013')],
+				'0123456789',
+				'0123',
+				'1234',
+			),
+			new Timestamps(
+				(new DateTime())->format('Y-m-d'),
+				(new DateTime('+ 1 week'))->format('Y-m-d'),
+			),
+		);
+
+		self::addItems($order, $itemCount, ':value Kč');
+
+		return $order;
+	}
+
+	private static function addItems(IOrder $order, ?int $itemCount = null, string $format = '$ :value'): void
+	{
+		$format = fn (string $money) => str_replace(':value', $money, $format);
+		$items = [
+			new Item('Logitech G700s Rechargeable Gaming Mouse', $format('1.790,00'), 4, $format('7.160,00')),
+			new Item('ASUS Z380KL 8" - 16GB, LTE', $format('6.490,00'), 1, $format('6.490,00')),
+			new Item('Philips 48PFS6909 - 121cm', $format('13.990,00'), 1, $format('13.990,00')),
+			new Item('HP Deskjet 3545 Advantage', $format('1.799,00'), 1, $format('1.799,00')),
+			new Item('LG 105UC9V - 266cm', $format('11.599,00'), 1, $format('11.599,00')),
+			new Item('Samsung Galaxy S21 Ultra 5G, 12GB/128GB', $format('31.490,00'), 1, $format('31.490,00')),
+		];
+		$count = count($items);
+
+		if ($itemCount === null) {
+			$itemCount = $count;
+		} else {
+			$itemCount = max($itemCount, 0);
+		}
+
+		if ($count < $itemCount) {
+			$array = $items;
+			for ($i = 0; $i < $itemCount - $count; $i++) {
+				$array[] = $items[array_rand($items)];
+			}
+
+			$items = $array;
+		} else {
+			$items = array_slice($items, 0, $itemCount);
+		}
+
+		foreach ($items as $item) {
+			$order->addItem($item);
+		}
 	}
 
 }
