@@ -1,17 +1,26 @@
 <?php declare(strict_types = 1);
 
-namespace Contributte\Invoice\Data;
+namespace Contributte\Invoice\Provider;
 
-class PaymentInformation implements IPaymentInformation
+use Contributte\Invoice\Data\IAccount;
+use Contributte\Invoice\Data\ICompany;
+
+final class InvoiceDataProvider
 {
 
 	/**
 	 * @param IAccount[] $accounts
 	 */
 	public function __construct(
+		private ICompany $company,
 		private array $accounts = [],
 	)
 	{
+	}
+
+	public function getCompany(): ICompany
+	{
+		return $this->company;
 	}
 
 	/**
@@ -20,11 +29,6 @@ class PaymentInformation implements IPaymentInformation
 	public function getAccounts(): array
 	{
 		return $this->accounts;
-	}
-
-	public function getFirstAccount(): ?IAccount
-	{
-		return $this->accounts[array_key_first($this->accounts)] ?? null;
 	}
 
 }
